Primary should not self-contain unless there's label or modifier
The generators in neotoma_qc were making nested primaries that didn't
compare equivalent to the round-tripped version, but I also realized it
was possible to parse in some very simple cases of nested parentheses
without modifiers. It might be harmless, but it's also useless.

The pretty-printer should also properly escape binaries as well as
lists.